Split out UCItemPricer into and interface. Implemented a dynamic pricing version.
