+++ build.gradle
@@ -101,7 +101,7 @@ sourceSets {
 
 
 tasks.withType(JavaCompile) {
-    sourceCompatibility = "1.7"
-    targetCompatibility = "1.7"
+    sourceCompatibility = "1.8"
+    targetCompatibility = "1.8"
     //options.compilerArgs << "-source 7"
 }
