+++ java/universalcoins/commands/UCCommand.java
@@ -10,13 +10,13 @@ import net.minecraft.command.ICommand;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.command.WrongUsageException;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.translation.I18n;
-import universalcoins.util.UCItemPricer;
+import universalcoins.UniversalCoins;
+import universalcoins.util.UCStaticItemPricer;
 
 public class UCCommand extends CommandBase implements ICommand {
 
@@ -59,7 +59,7 @@ public class UCCommand extends CommandBase implements ICommand {
 					new TextComponentString(I18n.translateToLocal("command.uccommand.option.update.help")));
 			sender.addChatMessage(new TextComponentString(I18n.translateToLocal("command.uccommand.usage.hint")));
 		} else if (args[0].matches(I18n.translateToLocal("command.uccommand.option.reload.name"))) {
-			UCItemPricer.getInstance().loadConfigs();
+			UniversalCoins.itemPricer.loadConfigs();
 		} else if (args[0].matches(I18n.translateToLocal("command.uccommand.option.get.name"))) {
 			// get item price
 			if (args.length > 1) {
@@ -68,20 +68,20 @@ public class UCCommand extends CommandBase implements ICommand {
 				if (args[1].matches(I18n.translateToLocal("command.uccommand.option.set.itemheld"))) {
 					ItemStack stack = getPlayerItem(sender);
 					if (stack != null) {
-						price = UCItemPricer.getInstance().getItemPrice(stack);
+						price = UniversalCoins.itemPricer.getItemPrice(stack);
 						stackName = getPlayerItem(sender).getDisplayName();
 					}
 				}
 				if (price == -1) {
 					sender.addChatMessage(new TextComponentString(
-							"§c" + I18n.translateToLocal("command.uccommand.warning.pricenotset") + " " + stackName));
+							"ï¿½c" + I18n.translateToLocal("command.uccommand.warning.pricenotset") + " " + stackName));
 				} else
 					sender.addChatMessage(
-							new TextComponentString("§a" + I18n.translateToLocal("command.uccommand.warning.pricefound")
+							new TextComponentString("ï¿½a" + I18n.translateToLocal("command.uccommand.warning.pricefound")
 									+ " " + stackName + ": " + formatter.format(price)));
 			} else
 				sender.addChatMessage(
-						new TextComponentString("§c" + I18n.translateToLocal("command.uccommand.warning.noitem")));
+						new TextComponentString("ï¿½c" + I18n.translateToLocal("command.uccommand.warning.noitem")));
 		} else if (args[0].matches(I18n.translateToLocal("command.uccommand.option.set.name"))) {
 			// set item price
 			if (args.length > 2) {
@@ -91,13 +91,13 @@ public class UCCommand extends CommandBase implements ICommand {
 					price = Integer.parseInt(args[2]);
 				} catch (NumberFormatException e) {
 					sender.addChatMessage(new TextComponentString(
-							"§c" + I18n.translateToLocal("command.uccommand.option.set.price.invalid")));
+							"ï¿½c" + I18n.translateToLocal("command.uccommand.option.set.price.invalid")));
 					return;
 				}
 				if (args[1].matches(I18n.translateToLocal("command.uccommand.option.set.itemheld"))) {
 					ItemStack stack = getPlayerItem(sender);
 					if (stack != null) {
-						result = UCItemPricer.getInstance().setItemPrice(stack, price);
+						result = UniversalCoins.itemPricer.setItemPrice(stack, price);
 					}
 				}
 				if (result == true) {
@@ -114,27 +114,27 @@ public class UCCommand extends CommandBase implements ICommand {
 					}
 				} else {
 					sender.addChatMessage(new TextComponentString(
-							"§c" + I18n.translateToLocal("command.uccommand.option.set.price.fail.one")));
+							"ï¿½c" + I18n.translateToLocal("command.uccommand.option.set.price.fail.one")));
 				}
 			} else
 				sender.addChatMessage(new TextComponentString(
-						"§c" + I18n.translateToLocal("command.uccommand.option.set.price.error")));
+						"ï¿½c" + I18n.translateToLocal("command.uccommand.option.set.price.error")));
 		} else if (args[0].matches(I18n.translateToLocal("command.uccommand.option.reload"))) {
-			UCItemPricer.getInstance().loadConfigs();
+			UniversalCoins.itemPricer.loadConfigs();
 			sender.addChatMessage(
-					new TextComponentString("§a" + I18n.translateToLocal("command.uccommand.option.reload.confirm")));
+					new TextComponentString("ï¿½a" + I18n.translateToLocal("command.uccommand.option.reload.confirm")));
 		} else if (args[0].matches(I18n.translateToLocal("command.uccommand.option.reset.name"))) {
-			UCItemPricer.getInstance().resetDefaults();
+			UniversalCoins.itemPricer.resetDefaults();
 			sender.addChatMessage(
-					new TextComponentString("§a" + I18n.translateToLocal("command.uccommand.option.reset.confirm")));
+					new TextComponentString("ï¿½a" + I18n.translateToLocal("command.uccommand.option.reset.confirm")));
 		} else if (args[0].matches(I18n.translateToLocal("command.uccommand.option.save.name"))) {
-			UCItemPricer.getInstance().savePriceLists();
+			UniversalCoins.itemPricer.savePriceLists();
 			sender.addChatMessage(
-					new TextComponentString("§a" + I18n.translateToLocal("command.uccommand.option.save.confirm")));
+					new TextComponentString("ï¿½a" + I18n.translateToLocal("command.uccommand.option.save.confirm")));
 		} else if (args[0].matches(I18n.translateToLocal("command.uccommand.option.update.name"))) {
-			UCItemPricer.getInstance().updatePriceLists();
+			UniversalCoins.itemPricer.updatePriceLists();
 			sender.addChatMessage(
-					new TextComponentString("§a" + I18n.translateToLocal("command.uccommand.option.update.confirm")));
+					new TextComponentString("ï¿½a" + I18n.translateToLocal("command.uccommand.option.update.confirm")));
 		}
 	}
 
@@ -166,7 +166,7 @@ public class UCCommand extends CommandBase implements ICommand {
 					|| args[0].matches(I18n.translateToLocal("command.uccommand.option.set.name"))) {
 				List<String> options = new ArrayList<String>();
 				options.add(I18n.translateToLocal("command.uccommand.option.set.itemheld"));
-				for (String item : UCItemPricer.getInstance().getUcPriceMap().keySet()) {
+				for (String item : UniversalCoins.itemPricer.getPricedItems()) {
 					options.add(item);
 				}
 				return getListOfStringsMatchingLastWord(args, options);
